+++ src/modules/plugins/custom-plugins/custom-plugins.module.ts
@@ -1,9 +1,13 @@
 import { Module } from '@nestjs/common';
 import { HomebridgeRingModule } from './homebridge-ring/homebridge-ring.module';
+import { HomebridgeHueModule } from './homebridge-hue/homebridge-hue.module';
 
 @Module({
   imports: [
     HomebridgeRingModule,
+    HomebridgeHueModule,
   ],
+  controllers: [],
+  providers: [],
 })
 export class CustomPluginsModule { }